feat: добавить регистрацию

+++ controllers/cards.js
@@ -1,82 +1,56 @@
 const Card = require('../models/card');
 const HttpError = require('../utils/HttpError');
 
-const DEFAULT_ERROR = 500;
-const VALIDATION_ERROR = 400;
-const NOT_FOUND_ERROR = 404;
-
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .populate(['owner', 'likes'])
     .then((cards) => res.send(cards))
-    .catch((err) => res.status(DEFAULT_ERROR).send({ message: err.message }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.send(card))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        res.status(VALIDATION_ERROR).send({ message: 'Переданы некорректные данные при создании карточки' });
-        return;
-      }
-      res.status(DEFAULT_ERROR).send({ message: err.message });
-    });
+    .catch(next);
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(new HttpError('idError', 'Карточка с указанным _id не найдена'))
-    .populate(['owner', 'likes'])
-    .then((card) => res.send(card))
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(VALIDATION_ERROR).send({ message: 'Переданы некорректные данные для удаления карточки' });
-      } else if (err.name === 'idError') {
-        res.status(NOT_FOUND_ERROR).send({ message: 'Карточка с указанным _id не найдена' });
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .orFail(new HttpError(404, 'idError', 'Карточка по указанному id не найдена'))
+    .then((card) => {
+      if (card.owner.toString() === req.user._id.toString()) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .populate(['owner', 'likes'])
+          .then((delCard) => res.send(delCard))
+          .catch(next);
       } else {
-        res.status(DEFAULT_ERROR).send({ message: err.message });
+        throw new HttpError(403, 'forbiddenError', 'Нет прав для удаления');
       }
-    });
+    })
+    .catch(next);
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
     { new: true },
   )
-    .orFail(new HttpError('idError', 'Карточка с указанным _id не найдена'))
+    .orFail(new HttpError(404, 'idError', 'Карточка по указанному id не найдена'))
     .populate(['owner', 'likes'])
     .then((card) => res.send(card))
-    .catch((err) => {
-      if (err.name === 'ValidationError' || err.name === 'CastError') {
-        res.status(VALIDATION_ERROR).send({ message: 'Переданы некорректные данные для постановки лайка' });
-      } else if (err.name === 'idError') {
-        res.status(NOT_FOUND_ERROR).send({ message: 'Карточка с указанным _id не найдена' });
-      } else {
-        res.status(DEFAULT_ERROR).send({ message: err.message });
-      }
-    });
+    .catch(next);
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
     { new: true },
   )
-    .orFail(new HttpError('idError', 'Карточка с указанным _id не найдена'))
+    .orFail(new HttpError(404, 'idError', 'Карточка по указанному id не найдена'))
     .populate(['owner', 'likes'])
     .then((card) => res.send(card))
-    .catch((err) => {
-      if (err.name === 'ValidationError' || err.name === 'CastError') {
-        res.status(VALIDATION_ERROR).send({ message: 'Переданы некорректные данные для снятия лайка' });
-      } else if (err.name === 'idError') {
-        res.status(NOT_FOUND_ERROR).send({ message: 'Карточка с указанным _id не найдена' });
-      } else {
-        res.status(DEFAULT_ERROR).send({ message: err.message });
-      }
-    });
+    .catch(next);
 };
